web_demo/requirements.txt
gradio>=5.0.0
ollama>=0.6.1
pillow>=12.2.0
pdf2image>=1.17.0
numpy>=1.26.0
scipy>=1.12.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.18
huggingface_hub>=0.26.0
